pyproject.toml
[project]
name = "uaarn-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deep-translator>=1.11.4",
    "docx>=0.2.4",
    "fastapi>=0.119.0",
    "gtts>=2.5.4",
    "langdetect>=1.0.9",
    "openai-agents[litellm]>=0.3.3",
    "pdfminer-six>=20250506",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.4",
    "tiktoken>=0.12.0",
    "uvicorn>=0.37.0",
    "youtube-transcript-api>=1.2.3",
]
